openpilot v0.9.7 release
date: 2024-03-24T10:14:01
master commit: 82a07afb2111b4367742b97ef7073b441d371891

+++ panda/board/drivers/usb.h
@@ -364,7 +364,7 @@ int current_int0_alt_setting = 0;
 
 void *USB_ReadPacket(void *dest, uint16_t len) {
   uint32_t *dest_copy = (uint32_t *)dest;
-  uint32_t count32b = (len + 3U) / 4U;
+  uint32_t count32b = ((uint32_t)len + 3U) / 4U;
 
   for (uint32_t i = 0; i < count32b; i++) {
     *dest_copy = USBx_DFIFO(0);
@@ -379,9 +379,9 @@ void USB_WritePacket(const void *src, uint16_t len, uint32_t ep) {
   hexdump(src, len);
   #endif
 
-  uint32_t numpacket = (len + (USBPACKET_MAX_SIZE - 1U)) / USBPACKET_MAX_SIZE;
+  uint32_t numpacket = ((uint32_t)len + (USBPACKET_MAX_SIZE - 1U)) / USBPACKET_MAX_SIZE;
   uint32_t count32b = 0;
-  count32b = (len + 3U) / 4U;
+  count32b = ((uint32_t)len + 3U) / 4U;
 
   // TODO: revisit this
   USBx_INEP(ep)->DIEPTSIZ = ((numpacket << 19) & USB_OTG_DIEPTSIZ_PKTCNT) |
